Switch to grid-based neighbor search and parallelize
This changes the neighbor search to use a grid-based indexing scheme
instead of the nanoflann-based kd-tree. We reorder the input points into
eps^2-sized cells. This way we can directly access the points of such a
cell and only have to query the 8-neighborhood of a point's cell to find
all its potential neighbors. For most datasets this is faster than first
building up the kd-tree and then querying it.

We further change the algorithm to require only a single pass over the
data, where for each point we find its eps-neighborhood and record if
that point could be a potential core point for any point in the
eps-neighborhood. We remember up two three potential core points. In the
next step, for each point that has at least one potential core point, we
label that point with its index (it for sure will be part of a cluster).
We finally need to merge the clusters iteratively until convergence in
the last step.

Co-authored-by: Alexander Hans <ahans@users.noreply.github.com>